+++ modules/AWSPowerShell/Cmdlets/ConnectCases/Basic/New-CCASDomain-Cmdlet.cs
@@ -34,7 +34,9 @@ namespace Amazon.PowerShell.Cmdlets.CCAS
     ///  <important><para>
     /// This will not associate your connect instance to Cases domain. Instead, use the Amazon
     /// Connect <a href="https://docs.aws.amazon.com/connect/latest/APIReference/API_CreateIntegrationAssociation.html">CreateIntegrationAssociation</a>
-    /// API.
+    /// API. You need specific IAM permissions to successfully associate the Cases domain.
+    /// For more information, see <a href="https://docs.aws.amazon.com/connect/latest/adminguide/required-permissions-iam-cases.html#onboard-cases-iam">Onboard
+    /// to Cases</a>.
     /// </para></important>
     /// </summary>
     [Cmdlet("New", "CCASDomain", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]