+++ modules/AWSPowerShell/Cmdlets/Redshift/Basic/Copy-RSClusterSnapshot-Cmdlet.cs
@@ -68,8 +68,8 @@ namespace Amazon.PowerShell.Cmdlets.RS
         /// <summary>
         /// <para>
         /// <para>The identifier of the cluster the source snapshot was created from. This parameter
-        /// is required if your IAM user has a policy containing a snapshot resource element that
-        /// specifies anything other than * for the cluster name.</para><para>Constraints:</para><ul><li><para>Must be the identifier for a valid cluster.</para></li></ul>
+        /// is required if your IAM user or role has a policy containing a snapshot resource element
+        /// that specifies anything other than * for the cluster name.</para><para>Constraints:</para><ul><li><para>Must be the identifier for a valid cluster.</para></li></ul>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]