+++ modules/AWSPowerShell/Cmdlets/SecurityHub/AWS.Tools.SecurityHub.Completers.psm1
@@ -87,6 +87,16 @@ $SHUB_Completers = {
             break
         }
 
+        # Amazon.SecurityHub.ControlFindingGenerator
+        {
+            ($_ -eq "Enable-SHUBSecurityHub/ControlFindingGenerator") -Or
+            ($_ -eq "Update-SHUBSecurityHubConfiguration/ControlFindingGenerator")
+        }
+        {
+            $v = "SECURITY_CONTROL","STANDARD_CONTROL"
+            break
+        }
+
         # Amazon.SecurityHub.ControlStatus
         "Update-SHUBStandardsControl/ControlStatus"
         {
@@ -132,6 +142,7 @@ $SHUB_Completers = {
 
 $SHUB_map = @{
     "AutoEnableStandards"=@("Update-SHUBOrganizationConfiguration")
+    "ControlFindingGenerator"=@("Enable-SHUBSecurityHub","Update-SHUBSecurityHubConfiguration")
     "ControlStatus"=@("Update-SHUBStandardsControl")
     "RecordState"=@("Update-SHUBFinding")
     "Severity_Label"=@("Update-SHUBFindingsBatch")
@@ -193,8 +204,11 @@ $SHUB_SelectMap = @{
                "Confirm-SHUBInvitation",
                "Disable-SHUBStandardsBatch",
                "Enable-SHUBStandardsBatch",
+               "Get-SHUBGetSecurityControl",
+               "Get-SHUBGetStandardsControlAssociation",
                "Import-SHUBFindingsBatch",
                "Update-SHUBFindingsBatch",
+               "Edit-SHUBUpdateStandardsControlAssociation",
                "New-SHUBActionTarget",
                "New-SHUBFindingAggregator",
                "New-SHUBInsight",
@@ -235,6 +249,8 @@ $SHUB_SelectMap = @{
                "Get-SHUBInvitationList",
                "Get-SHUBMemberList",
                "Get-SHUBOrganizationAdminAccountList",
+               "Get-SHUBSecurityControlDefinitionList",
+               "Get-SHUBStandardsControlAssociationList",
                "Get-SHUBResourceTag",
                "Add-SHUBResourceTag",
                "Remove-SHUBResourceTag",