+++ modules/AWSPowerShell/Cmdlets/ServiceCatalog/Basic/Get-SCProduct-Cmdlet.cs
@@ -29,6 +29,11 @@ namespace Amazon.PowerShell.Cmdlets.SC
 {
     /// <summary>
     /// Gets information about the specified product.
+    /// 
+    ///  <note><para>
+    ///  Running this operation with administrator access results in a failure. <a>DescribeProductAsAdmin</a>
+    /// should be used instead. 
+    /// </para></note>
     /// </summary>
     [Cmdlet("Get", "SCProduct")]
     [OutputType("Amazon.ServiceCatalog.Model.DescribeProductResponse")]